\name{cat_function}
\alias{cat_function}
\title{
cat function
}
\description{
This function prints a cute cat picture if you are a cat lover.
}
\usage{
cat_function(love = TRUE)
}
\arguments{
  \item{love}{
a Boolean TRUE/FALSE value indicating whether you love cats (TRUE) or do not love cats (FALSE)
}
}
\details{
}
\value{
Prints a character string (which looks like a cat).
}
\references{
}
\author{
Joyce Njuguna & Marlee Labroo
}
\note{
}


\seealso{

}
\examples{
"Here is an example of the execution should you love cats:
cat_function(love = TRUE)
Here is an example of the execution should you not love cats for some reason:
cat_function(love = FALSE)"
}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{}
\keyword{}
